pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bls-sdk"
version = "0.1.0"
description = "Python client for the Bureau of Labor Statistics APIs"
authors = [{name = "you"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
	"requests>=2.32.3",
	"python-dotenv>=1.0.1",
	"tenacity>=9.0.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["bls_sdk*"]
exclude = ["tests*"]
